+++ pkg/provider/dataset_from_file_resource.go
@@ -163,7 +163,7 @@ func (r *DatasetFromFileResource) Read(ctx context.Context, req resource.ReadReq
 			resp.State.RemoveResource(ctx)
 		} else {
 			resp.Diagnostics.AddError(
-				fmt.Sprintf("Error getting Dataset with ID %s", data.ID.ValueString()), 
+				fmt.Sprintf("Error getting Dataset with ID %s", data.ID.ValueString()),
 				err.Error())
 		}
 		return