+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.18 (Unreleased)
+## 0.0.19
 
 FEATURES:
+
+- Rename `chat_application` to `qa_application`
+- Support other types of Runtime Parameters besides string (type conversion is handled internally).
+- Abort Deployment create if an Error occurs